--- middleware/validateDeveloper.js
@@ -1,55 +1,0 @@
-function validateEmail(email) {
-    return String(email)
-        .toLowerCase()
-        .match(
-            /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-        );
-};
-exports.validDevInput = (req, res,next) => {
-    const {name,phone,domain,email,password} = req.body;
-
-    function ValidURL(str) {
-        var regex = /(?:https?):\/\/(\w+:?\w*)?(\S+)(:\d+)?(\/|\/([\w#!:.?+=&%!\-\/]))?/;
-        if (!regex.test(str)) {
-            let inputError = "please enter a proper domain name with https protcol";
-            res.render('login&signup/developer-register', {
-                Ierror: inputError
-            })
-            return
-        }
-        return
-    }
-
-    if (name.length < 3) {
-        let inputError = "please enter a valid name";
-        res.render('login&signup/developer-register', {
-            Ierror: inputError
-        })
-        return
-    }
-    if (password.length < 8) {
-        let inputError = "please use a password that have more than 8 characters";
-        res.render('login&signup/developer-register', {
-            Ierror: inputError
-        })
-        return
-    }
-    if (typeof phone == 'number') {
-        let inputError = "please enter a valid phone";
-        res.render('login&signup/developer-register', {
-            Ierror: inputError
-        })
-        return
-    }
-
-    ValidURL(domain)
-
-    if (!validateEmail(email)) {
-        let inputError = "please enter a valid Email Address";
-        res.render('login&signup/developer-register', {
-            Ierror: inputError
-        })
-        return
-    }
-    next()
-}
